Refactor av mappenavn til "oppgaveliste"
