+++ src/felles-komponenter/forside/oppgaveliste/journalForingEnkeltLinje.less
