+++ src/felles-komponenter/forside/oppgaveliste/sakEnkeltLinje.less
